requirements.txt
Django==5.1.2
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
firebase-admin==6.5.0